[1-6] htmlStatement 추가
- renderHtml 정의
- html 테스트 추가
